+++ StabilityMatrix.Avalonia/DesignData/DesignData.cs
@@ -404,7 +404,8 @@ public static class DesignData
                     new PackageModificationRunner
                     {
                         CurrentProgress = new ProgressReport(0.5f, "Installing package...")
-                    }
+                    },
+                    new List<IPackageStep>()
                 )
             }
         );
@@ -444,6 +445,52 @@ public static class DesignData
     public static HuggingFacePageViewModel HuggingFacePageViewModel =>
         Services.GetRequiredService<HuggingFacePageViewModel>();
 
+    public static NewOneClickInstallViewModel NewOneClickInstallViewModel =>
+        Services.GetRequiredService<NewOneClickInstallViewModel>();
+
+    public static RecommendedModelsViewModel RecommendedModelsViewModel =>
+        DialogFactory.Get<RecommendedModelsViewModel>(vm =>
+        {
+            vm.Sd15Models = new ObservableCollectionExtended<RecommendedModelItemViewModel>()
+            {
+                new()
+                {
+                    ModelVersion = new CivitModelVersion
+                    {
+                        Name = "BB95 Furry Mix",
+                        Description = "A furry mix of BB95",
+                        Stats = new CivitModelStats { Rating = 3.5, RatingCount = 24 },
+                        Images =
+                        [
+                            new CivitImage
+                            {
+                                Url =
+                                    "https://image.civitai.com/xG1nkqKTMzGDvpLrqFT7WA/78fd2a0a-42b6-42b0-9815-81cb11bb3d05/00009-2423234823.jpeg"
+                            }
+                        ],
+                    },
+                    Author = "bb95"
+                },
+                new()
+                {
+                    ModelVersion = new CivitModelVersion
+                    {
+                        Name = "BB95 Furry Mix",
+                        Description = "A furry mix of BB95",
+                        Stats = new CivitModelStats { Rating = 3.5, RatingCount = 24 },
+                        Images =
+                        [
+                            new CivitImage
+                            {
+                                Url =
+                                    "https://image.civitai.com/xG1nkqKTMzGDvpLrqFT7WA/78fd2a0a-42b6-42b0-9815-81cb11bb3d05/00009-2423234823.jpeg"
+                            }
+                        ],
+                    },
+                    Author = "bb95"
+                }
+            };
+        });
     public static OutputsPageViewModel OutputsPageViewModel
     {
         get
@@ -490,6 +537,14 @@ public static class DesignData
         {
             vm.AddExtensions(
                 [
+                    new PackageExtension
+                    {
+                        Author = "123",
+                        Title = "Cool Extension",
+                        Description = "This is an interesting extension",
+                        Reference = new Uri("https://github.com/LykosAI/StabilityMatrix"),
+                        Files = [new Uri("https://github.com/LykosAI/StabilityMatrix")]
+                    },
                     new PackageExtension
                     {
                         Author = "123",