+++ StabilityMatrix.Core/Api/ICivitApi.cs
@@ -7,10 +7,13 @@ public interface ICivitApi
 {
     [Get("/api/v1/models")]
     Task<CivitModelsResponse> GetModels(CivitModelsRequest request);
-    
+
     [Get("/api/v1/models/{id}")]
     Task<CivitModel> GetModelById([AliasAs("id")] int id);
-    
+
     [Get("/api/v1/model-versions/by-hash/{hash}")]
     Task<CivitModelVersionResponse> GetModelVersionByHash([Query] string hash);
+
+    [Get("/api/v1/model-versions/{id}")]
+    Task<CivitModelVersion> GetModelVersionById(int id);
 }